tavily-news-agent/requirements.txt
tavily-python
anthropic[bedrock]
python-dotenv
duckduckgo-search
